1. Updated README.md for more clarity. 2. Renamed variables for more clarity. 3. Moved files around for more clarity
