+++ DeepLabCut/bash_scripts/Windows/Win_label_frames.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-cd ../../whiskers/
+cd ../../code/
 echo "activating environment"
 conda activate dlc-windowsGPU
-python add_videos.py
+python label.py
 read -p "all set: press enter to continue"
