+++ DeepLabCut/bash_scripts/Windows/dlc-windowsCPU.yaml
