+++ DeepLabCut/bash_scripts/check_os.sh
