+++ DeepLabCut/bash_scripts/dlc-windowsGPU.yaml
