+++ DeepLabCut/code/check_os.py
