+++ DeepLabCut/code/config.json
@@ -1,0 +1,3 @@
+{
+    "path_config_file": "C:\\Users\\Zack\\Documents\\GitHub\\neuro-circuit\\DeepLabCut\\test-task-zacklit-2019-06-21\\config.yaml"
+}
