+++ DeepLabCut/code/fragments.py
