+++ DeepLabCut/pre_config.yaml
@@ -2,26 +2,29 @@
 # Required
 # If the videos are in a sub-folder of the current folder, just type ./ and the name of the folder
 # Feel free to change the following example (./raw_videos)
-relative_Path_to_videos: ./raw_videos
+relative_path_to_videos: ../raw_videos
 bodyparts:
   - whisker_1
   - whisker_2
+  - whisker_3
+  - whisker_4
+  - else
 
 # Optional
 task:
 researchers:
 # Change the working directory to where you want the folders created. If left blank, the current directory is used
-working_directory:
+relative_path_to_working_directory:
 frame_extraction_mode:
 frame_extraction_algorithm:
 # The number of frames to pick from a video
-numframes2pick: 25
+numframes2pick:
 
 # Default Options
 default_values:
   task: whiskers
   researchers: the_Chen_Lab
-  working_directory: current directory
+  relative_path_to_working_directory: ../
   frame_extraction_mode: automatic
   frame_extraction_algorithm: kmeans
   numframes2pick: 20
