--- bash_scripts/Mac/extract_frames_DLC_Mac.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-# move to /c/Users/YOUR_USERNAME/Documents/GitHub/neuro-circuit/bash_scripts
-echo "activating environment"
-source activate test-dlc-macOS-CPU
-python run.py
-read -p "all set: press enter to continue"
