--- whiskers/config.json
@@ -1,3 +1,0 @@
-{
-    "path_config_file": "C:\\Users\\Zack\\Documents\\GitHub\\neuro-circuit\\whiskers\\whiskers-the_Chen_Lab-2019-06-19\\config.yaml"
-}
